+++ src/NavigationBar/tests/OffsetNavPush.test.js
@@ -2,28 +2,45 @@
  * Testing our OffsetNavPush component
  */
 
-import { mount } from 'enzyme';
+import { shallow, mount } from 'enzyme';
 import React from 'react';
-import { BootstrapProvider, makeTheme } from 'bootstrap-styled';
-import { theme } from '../theme';
+import { BootstrapProvider } from 'bootstrap-styled';
+import NavigationBar, { defaultProps } from '../NavigationBar';
 import OffsetNavPush from '../OffsetNavPush';
 
+const children = (<h1>Test</h1>);
+
 describe('<OffsetNavPush />', () => {
+  let theme;
   let props;
 
   beforeEach(() => {
-    props = {
-      theme: makeTheme(theme),
-      children: <h1>Test</h1>,
-    };
+    theme = defaultProps.theme;
+    props = Object.assign(defaultProps, {
+      children,
+    });
   });
 
   it('should render an OffsetNavPush', () => {
-    const renderedComponent = mount(
-      <BootstrapProvider theme={props.theme}>
-        <OffsetNavPush {...props} />
+    const renderedComponent = shallow(
+      <BootstrapProvider theme={theme}>
+        <NavigationBar {...props}>
+          <OffsetNavPush>
+            {children}
+          </OffsetNavPush>
+        </NavigationBar>
       </BootstrapProvider>
     );
     expect(renderedComponent.find('OffsetNavPush').length).toEqual(1);
+  });
+  it('should render an OffsetNavPush with show xs', () => {
+    const renderedComponent = mount(
+      <BootstrapProvider theme={theme}>
+        <NavigationBar {...Object.assign(props, { offsetNav: { show: 'xs', push: true } })} >
+          {children}
+        </NavigationBar>
+      </BootstrapProvider>
+    );
+    expect(renderedComponent.find('OffsetNavPush').props().show).toEqual('xs');
   });
 });